Support autodoxygenfile directive
The initial index.py refactoring was based on the assumption that we
couldn't share a lot of the node and transform code. It feels like we
can for the moment so we've moved some of the index.py code back into
more generic locations and names in transforms.py, nodes.py and base.py.

We then introduce file.py with the necessary directives and FileHandler
to do all the file specific work and then we adjust the transforms to
handle the Handlers and their data more generically.

The FileHanlder ends up being a cut-paste-and-tweak of the original
DoxygenFileDirective code which is promising for future changes.